+++ ceammc/ext/src/flow/flow_demultiplex.h
@@ -7,16 +7,20 @@ using namespace ceammc;
 
 class FlowDemultiplex : public BaseObject {
     SizeTProperty* index_;
+    FlagProperty* no_props_;
 
 public:
     FlowDemultiplex(const PdArgs& a);
 
-    void onBang();
-    void onFloat(t_float f);
-    void onSymbol(t_symbol* s);
-    void onList(const AtomList& l);
-    void onAny(t_symbol* s, const AtomList& l);
-    void onData(const DataPtr& d);
+    void onBang() override;
+    void onFloat(t_float f) override;
+    void onSymbol(t_symbol* s) override;
+    void onList(const AtomList& l) override;
+    void onAny(t_symbol* s, const AtomList& l) override;
+    void onData(const DataPtr& d) override;
+    void onInlet(size_t n, const AtomList& l) override;
+
+    bool processAnyProps(t_symbol* sel, const AtomList& lst) override;
 
     bool checkIndex() const;
 };